requirements.txt
python-telegram-bot==20.6
aiohttp
sqlalchemy>=2.0
asyncpg
python-dotenv
psycopg2-binary
